+++ Makefile
@@ -3,8 +3,8 @@ CXX      := c++
 CXXFLAGS := -Wall -Wextra -Werror -std=c++98 -MMD -MP
 SRCDIR   := srcs
 OBJDIR   := objs
-SRCS     := $(SRCDIR)/main.cpp $(SRCDIR)/server/Server.cpp $(SRCDIR)/utils.cpp \
-            $(SRCDIR)/config/config_parse.cpp \
+SRCS     := $(SRCDIR)/main.cpp $(SRCDIR)/server/Server.cpp $(SRCDIR)/Utils.cpp \
+            $(SRCDIR)/config/ConfigParse.cpp \
             $(SRCDIR)/event/EpollMultiplexer.cpp \
 			$(SRCDIR)/event/Multiplexer.cpp \
 			$(SRCDIR)/event/SelectMultiplexer.cpp \
