+++ includes/HttpResponse.hpp
@@ -6,7 +6,7 @@
 /*   By: koseki.yusuke <koseki.yusuke@student.42    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/02 16:44:51 by koseki.yusu       #+#    #+#             */
-/*   Updated: 2025/03/05 18:48:50 by koseki.yusu      ###   ########.fr       */
+/*   Updated: 2025/03/05 21:11:09 by koseki.yusu      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,5 +24,4 @@ public:
     static void send_error_response(int clientFd, int status_code, const std::string &message);
     static void send_response(int client_socket, int status_code, const std::string &content, const std::string &content_type);
     static void send_redirect(int client_socket, int status_code, const std::string new_location);
-    std::string read_file(const std::string& file_path);
 };
