+++ srcs/http/HttpResponse.cpp
@@ -6,16 +6,28 @@
 /*   By: koseki.yusuke <koseki.yusuke@student.42    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/02 16:37:08 by koseki.yusu       #+#    #+#             */
-/*   Updated: 2025/03/05 20:30:26 by koseki.yusu      ###   ########.fr       */
+/*   Updated: 2025/03/05 21:12:02 by koseki.yusu      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "HttpResponse.hpp"
 #include "HttpRequest.hpp"
+#include "Utils.hpp"
+
+void HttpResponse::send_response(int client_socket, int status_code, const std::string &content, const std::string &content_type) {
+    std::ostringstream response;
+    response << "HTTP/1.1 " << status_code << " OK\r\n";
+    response << "Content-Length: " << content.size() << "\r\n";
+    response << "Content-Type: " << content_type << "\r\n\r\n";
+    response << content;
+
+    send(client_socket, response.str().c_str(), response.str().size(), 0);
+}
+
 
 void HttpResponse::send_custom_error_page(int client_socket, int status_code, const std::string &error_page) {
     try {
-        std::string file_content = HttpResponse().read_file("./public/" + error_page);
+        std::string file_content = read_file("./public/" + error_page);
 
         std::ostringstream response;
         response << "HTTP/1.1 " << status_code << " ";
@@ -43,16 +55,6 @@ void HttpResponse::send_custom_error_page(int client_socket, int status_code, co
     }
 }
 
-std::string HttpResponse::read_file(const std::string& file_path) {
-    std::ifstream file(file_path.c_str(), std::ios::in);
-    if (!file.is_open()) {
-        throw std::runtime_error("Failed to open file: " + file_path);
-    }
-    std::ostringstream buffer;
-    buffer << file.rdbuf();
-    return buffer.str();
-}
-
 void HttpResponse::send_error_response(int client_socket, int status_code, const std::string &message) {
     std::ostringstream response;
     response << "HTTP/1.1 " << status_code << " " << message << "\r\n";
@@ -63,16 +65,7 @@ void HttpResponse::send_error_response(int client_socket, int status_code, const
     send(client_socket, response.str().c_str(), response.str().size(), 0);
 }
 
-void HttpResponse::send_response(int client_socket, int status_code, const std::string &content, const std::string &content_type) {
-    std::ostringstream response;
-    response << "HTTP/1.1 " << status_code << " OK\r\n";
-    response << "Content-Length: " << content.size() << "\r\n";
-    response << "Content-Type: " << content_type << "\r\n\r\n";
-    response << content;
-
-    send(client_socket, response.str().c_str(), response.str().size(), 0);
-}
-
+// 未実装
 void HttpResponse::send_redirect(int client_socket, int status_code, const std::string new_location)
 {
     std::cout << client_socket << std::endl;
